requirements.txt
setuptools>=69.0.0
pip>=24.0.0
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
pymongo==4.6.1
dnspython==2.6.1
schedule==1.2.1
python-dotenv==1.0.1
google-generativeai==0.3.2
werkzeug==2.3.7
sounddevice==0.4.6
numpy>=1.26.0
google-adk>=0.1.0
gunicorn==21.2.0 